refactor(sketching): consolidate SVG ellipse math into ellipseUtils
Layer 3 audit F3 — `sketcherlib.ts` mixed two unrelated concerns: the
public `GenericSketcher<T>` interface (~415 LOC) and SVG elliptical-arc
math helpers (`convertSvgEllipseParams`, `computeArcAngles`,
`radianAngle`, ~135 LOC). The math was already partially in
`ellipseUtils.ts` (`normalizeEllipseRadii`, `makeEllipseArcFromSvgParams`)
which then imported back from `sketcherlib`.

Move all three SVG functions into `ellipseUtils.ts` so all ellipse-arc
math lives in one file. `sketcherlib.ts` is now interface-only
(564 -> 415 LOC), `ellipseUtils.ts` is now self-contained (102 -> 245 LOC).

Behaviour unchanged.
